requirements.txt
torch=2.8.0
torchvision=0.23.0
pillow=11.3.0
numpy=2.3.3
